chore: Bump firebase to version 9 (modular). Refactor the code to leverage it.
I could use 'firebase/compat' imports and leave the code as it is, but I
decided to refactor it to leverage modular firebase. In addition to
that, I refactored leaked fire.js abstraction (moved all the logic to
fire.js and moved itself since it doesn't need a folder anymore).